reName appMessageService
