+++ src/services/pages/detail-service.ts
@@ -9,7 +9,7 @@ import { finalize } from "rxjs/operators";
 import { formService } from "services/FormService";
 import { routerService } from "services/RouterService";
 import { v4 as uuidv4 } from "uuid";
-import appMessageService from "services/AppMessageService";
+import appMessageService from "services/app-message-service";
 
 export class DetailService {
   /**
